Moves to new utils
