+++ main.py
@@ -10,6 +10,7 @@ from tqdm import tqdm
 from clip_loss import CLIPLoss
 from stylegan2.model import Generator
 import clip
+from utils import ensure_checkpoint_exists
 
 
 
@@ -19,26 +20,6 @@ def get_lr(t, initial_lr, rampdown=0.25, rampup=0.05):
     lr_ramp = lr_ramp * min(1, t / rampup)
 
     return initial_lr * lr_ramp
-
-
-def ensure_checkpoint_exists(model_weights_filename):
-    google_drive_paths = {
-        "stylegan2-ffhq-config-f.pt": "https://drive.google.com/uc?id=1EM87UquaoQmk17Q8d5kYIAHqu0dkYqdT"
-    }
-
-    if not os.path.isfile(model_weights_filename) and (model_weights_filename in google_drive_paths):
-        gdrive_url = google_drive_paths[model_weights_filename]
-        try:
-            from gdown import download as drive_download
-            drive_download(gdrive_url, model_weights_filename, quiet=False)
-        except ModuleNotFoundError:
-            print("""gdown not found.
-                pip3 install gdown or manually download the checkpoint file: 
-                """ + gdrive_url)
-
-    if not os.path.isfile(model_weights_filename) and (model_weights_filename not in google_drive_paths):
-        print(model_weights_filename +
-              ' not found, you may need to manually download the model weights.')
 
 
 def main(args):
